% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-keys.R
\name{element_send_keys}
\alias{element_send_keys}
\title{Send keys to element}
\usage{
element_send_keys(self, private, keys)
}
\arguments{
\item{self}{object}

\item{private}{private object}

\item{keys}{}
}
\description{
Send keys to element
}
\examples{
\dontrun{
el$send_keys("xyz")
el$send_keys("x", "y", "z")
el$send_keys("username", key$enter, "password", key$enter)

## Sending CTRL+A 
el$send_keys(key$control, "a")

## Note that modifier keys (control, alt, shift, meta) are sticky,
## they remain in effect in the rest of the send_keys() call. E.g.
## this sends CTRL+X and CTRL+S
el$send_keys(key$control, "x", "s")

## You will need multiple calls to release control and send CTRL+X S
el$send_keys(key$control, "x")
el$send_keys("s")
}
}
\keyword{internal}

