% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-keys.R
\name{element_sendKeys}
\alias{element_sendKeys}
\title{Send keys to element}
\usage{
element_sendKeys(self, private, ...)
}
\arguments{
\item{self}{object}

\item{private}{private object}

\item{...}{see examples at \code{\link{key}}}
}
\description{
Send keys to element
}
\keyword{internal}
