% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-chromedriver.R
\name{run_chromedriver}
\alias{run_chromedriver}
\title{Start up chromedriver on localhost, and a random port}
\usage{
run_chromedriver()
}
\value{
A list of \code{process}, the \code{processx::process} object,
  and \code{port}, the local port where chromedriver is running.
}
\description{
Throws an error if chromedriver cannot be found, or cannot be
started. It works with a timeout of five seconds.
}

