% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{webdriver}
\alias{webdriver}
\alias{webdriver-package}
\title{'WebDriver' Client for 'PhantomJS'}
\description{
A client for the 'WebDriver' 'API'. It allows driving a (probably
headless) web browser, and can be used to test web applications,
including 'Shiny' apps. In theory it works with any 'WebDriver'
implementation, but it was only tested with 'PhantomJS'.
}
