% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-keys.R
\docType{data}
\name{key}
\alias{key}
\title{Special keys, so that we can refer to them with an easier syntax}
\format{An object of class \code{list} of length 51.}
\usage{
key
}
\description{
Special keys, so that we can refer to them with an easier syntax
}
\examples{
\dontrun{
el$send_keys("xyz")
el$send_keys("x", "y", "z")
el$send_keys("username", keys$enter, "password", keys$enter)
}

}
\keyword{datasets}

