% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-phantomjs.R
\name{run_phantomjs}
\alias{run_phantomjs}
\title{Start up phantomjs on localhost, and a random port}
\usage{
run_phantomjs(debugLevel = c("INFO", "ERROR", "WARN", "DEBUG"),
  timeout = 5000)
}
\arguments{
\item{debugLevel}{Phantom.js debug level, possible values: \code{"INFO"},
\code{"ERROR"}, \code{"WARN"}, \code{"DEBUG"}.}

\item{timeout}{How long to wait (in milliseconds) for the webdriver
connection to be established to the phantomjs process.}
}
\value{
A list of \code{process}, the \code{callr::process} object, and
  \code{port}, the local port where phantom is running.
}
\description{
Throws and error if phantom cannot be found, or cannot be started. It works
with a timeout of five seconds.
}
